PWA-1282: Refactored determination of the scroll container within the InfiniteContainer
