+++ themes/theme-gmd/components/ProductList/index.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { ITEMS_PER_LOAD } from '@shopgate/pwa-common/constants/DisplayOptions';
 import InfiniteContainer from '@shopgate/pwa-common/components/InfiniteContainer';
 import LoadingIndicator from '@shopgate/pwa-ui-shared/LoadingIndicator';
+import { ViewContext } from 'Components/View/context';
 import Iterator from './components/Iterator';
 import Layout from './components/Layout';
 
@@ -34,17 +35,22 @@ const ProductList = ({
   }
 
   return (
-    <InfiniteContainer
-      wrapper={Layout}
-      iterator={Iterator}
-      loader={handleGetProducts}
-      items={products}
-      loadingIndicator={<LoadingIndicator />}
-      totalItems={totalProductCount}
-      initialLimit={10}
-      limit={ITEMS_PER_LOAD}
-      requestHash={requestHash}
-    />
+    <ViewContext.Consumer>
+      {({ getContentRef }) => (
+        <InfiniteContainer
+          containerRef={getContentRef()}
+          wrapper={Layout}
+          iterator={Iterator}
+          loader={handleGetProducts}
+          items={products}
+          loadingIndicator={<LoadingIndicator />}
+          totalItems={totalProductCount}
+          initialLimit={10}
+          limit={ITEMS_PER_LOAD}
+          requestHash={requestHash}
+        />
+      )}
+    </ViewContext.Consumer>
   );
 };
